Organize HSM.Benchmark's structure
